+++ tests/robots/test_robot.py
@@ -4,9 +4,11 @@ import re
 import pytest
 from compas.geometry import Frame
 from compas.robots import RobotModel
-
 from compas.robots.base_artist import BaseRobotModelArtist
 
+from compas_fab.backends.interfaces import ClientInterface
+from compas_fab.backends.interfaces import InverseKinematics
+from compas_fab.backends.interfaces import PlannerInterface
 from compas_fab.robots import Robot
 from compas_fab.robots import RobotSemantics
 from compas_fab.robots.ur5 import Robot as Ur5Robot
@@ -57,6 +59,20 @@ def ur5_joints(ur5_robot_instance):
 @pytest.fixture
 def ur5_links(ur5_robot_instance):
     return ur5_robot_instance.model.links
+
+
+@pytest.fixture
+def fake_client():
+    class FakeClient(ClientInterface):
+        pass
+
+    class FakePlanner(PlannerInterface):
+        pass
+
+    client = FakeClient()
+    client.planner = FakePlanner(client)
+
+    return client
 
 
 def test_basic_name_only():
@@ -210,3 +226,29 @@ def test_get_configurable_joints_wo_semantics(panda_robot_instance_wo_semantics)
     pattern = re.compile(r'panda_.*joint\d')
     matches = [pattern.match(joint.name) for joint in joints]
     assert all(matches)
+
+
+def test_inverse_kinematics_repeated_calls_will_return_next_result(ur5_robot_instance, fake_client):
+    class FakeInverseKinematics(InverseKinematics):
+        def inverse_kinematics(self, robot, frame_WCF, start_configuration=None, group=None, options=None):
+            results = [
+                ((-1.572, -2.560, 2.196, 2.365, 0.001, 1.137), (0, 0, 0, 0, 0, 0), ('shoulder_pan_joint', 'shoulder_lift_joint', 'elbow_joint', 'wrist_1_joint', 'wrist_2_joint', 'wrist_3_joint')),
+                ((-2.238, -3.175, 2.174, 4.143, -5.616, -6.283), (0, 0, 0, 0, 0, 0), ('shoulder_pan_joint', 'shoulder_lift_joint', 'elbow_joint', 'wrist_1_joint', 'wrist_2_joint', 'wrist_3_joint')),
+            ]
+
+            for ik in results:
+                yield (ik[0], ik[2])
+
+    robot = ur5_robot_instance
+    robot.client = fake_client
+    robot.client.inverse_kinematics = FakeInverseKinematics()
+
+    frame = Frame.worldXY()
+    start_config = robot.zero_configuration()
+
+    configuration = robot.inverse_kinematics(frame, start_config)
+    assert str(configuration) == "Configuration((-1.572, -2.560, 2.196, 2.365, 0.001, 1.137), (0, 0, 0, 0, 0, 0), ('shoulder_pan_joint', 'shoulder_lift_joint', 'elbow_joint', 'wrist_1_joint', 'wrist_2_joint', 'wrist_3_joint'))"
+    configuration = robot.inverse_kinematics(frame, start_config)
+    assert str(configuration) == "Configuration((-2.238, -3.175, 2.174, 4.143, -5.616, -6.283), (0, 0, 0, 0, 0, 0), ('shoulder_pan_joint', 'shoulder_lift_joint', 'elbow_joint', 'wrist_1_joint', 'wrist_2_joint', 'wrist_3_joint'))"
+    configuration = robot.inverse_kinematics(frame, start_config)
+    assert str(configuration) == "Configuration((-1.572, -2.560, 2.196, 2.365, 0.001, 1.137), (0, 0, 0, 0, 0, 0), ('shoulder_pan_joint', 'shoulder_lift_joint', 'elbow_joint', 'wrist_1_joint', 'wrist_2_joint', 'wrist_3_joint'))"