+++ CHANGELOG.rst
@@ -18,8 +18,8 @@ Unreleased
 
 **Changed**
 
-* Changed the backend feature `InverseKinematics.inverse_kinematics` to be a generator
-* Standardized the yielded type of `InverseKinematics.inverse_kinematics` across the PyBullet, MoveIt and V-REP planners
+* Changed the backend feature ``InverseKinematics.inverse_kinematics`` to be a generator. As a consequence of this, ``ClientInterface.inverse_kinematics`` and ``PlannerInterface.inverse_kinematics`` have changed to generators as well.
+* Standardized the yielded type of ``InverseKinematics.inverse_kinematics`` across the PyBullet, MoveIt and V-REP planners
 
 **Fixed**